+++ includes/corewar.h
@@ -6,7 +6,7 @@
 /*   By: matleroy <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/04 11:32:19 by matleroy          #+#    #+#             */
-/*   Updated: 2019/05/17 17:32:52 by acompagn         ###   ########.fr       */
+/*   Updated: 2019/05/18 00:31:25 by tle-dieu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,7 +63,7 @@ typedef	struct		s_champ
 	int				color;
 	int				id;
 	int				chosen_id[2];
-	ssize_t			size;
+	long			size;
 }					t_champ;
 
 typedef struct		s_visu
